finished removing my shame for the review, mostly deleteing comments and nmoving files
