+++ frontend/src/App2.js
@@ -1,5 +1,5 @@
 import React, { useEffect } from "react";
-import { useDispatch,useSelector } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { Switch, Route } from "react-router-dom";
 import * as sessionActions from "./store/session2";
 import Navigation from "./components2/Navigation";
@@ -9,21 +9,15 @@ import CurrentIndex from "./components2/CurrentIndex";
 import './index.css';
 
 
-//took out my 404 page made header and most of any excess
-//No routes working
-
 function App() {
     const dispatch = useDispatch();
     // const [isLoaded, setIsLoaded] = useState(false);
     let user = useSelector(state => state.session.user)
     useEffect(() => {
         // if(!user)
-        if(!user)dispatch(sessionActions.restoreUser())
-        // .then(() => dispatch(setIsLoaded(true)))
-        // setIsLoaded(true)
-        // .then(() => dispatch(setIsLoaded(true)))  ///took out again
+        if (!user) dispatch(sessionActions.restoreUser())
+        // .then(() => setIsLoaded(true))  ///took out again
     }, [dispatch]);
-    ///trying top 
     return (
 
         <>
@@ -45,9 +39,9 @@ function App() {
                     </Route>
 
                     <Route>
-                        <div style={{display: 'flex', flexDirection: 'column', alignItems: 'center'}}>
-                        <h1 style={{ paddingTop: '5vw', fontFamily: 'Bold'}}>Sorry! Out Of Luck!</h1>
-                        <h3 style={{ paddingTop: '8vw'}}>Sorry, we could not find what you are looking for</h3>
+                        <div style={{ display: 'flex', flexDirection: 'column', alignItems: 'center' }}>
+                            <h1 style={{ paddingTop: '5vw', fontFamily: 'Bold' }}>Sorry! Out Of Luck!</h1>
+                            <h3 style={{ paddingTop: '8vw' }}>Sorry, we could not find what you are looking for</h3>
                         </div>
                     </Route>
