+++ frontend/src/components2/CreateSpotImageModal/index.js
@@ -47,19 +47,7 @@ function SpotImageForm({ idx, spot }) {
                     <div id={!errors.length ? 'signupsubheader' : 'errorswap'}>{!errors.length ? `Welcome ${onespots[id]?.User?.firstName}!` : errors.map((error, idx) => <>{error}<br /></>)}</div>
                 </div>
             </div>
-            {/* <div id='createheadersub' className='spotimagehead'>
-                <div id='createexitbutt' onClick={() => closeModal()}>
-                    x
-                </div>
-                <div id='createheadertext' className='spotimagesubhead'>
-                    <div id="createmainheader">Welcome</div>
-                    <div id='createsubheader'>Add Photos to your Spot!</div>
-                </div>
-            </div> */}
 
-            {/* <ul id='errorsimage'>
-                {errors.map((error, idx) => <li className='errors' key={error + idx}>{error}</li>)}
-            </ul> */}
             <div id='createpreviewcont'>
                 <div id='previmgwrap1'>
                     Current Preview Image
