+++ frontend/src/components2/EditReviewForm/index.js
@@ -1,19 +1,11 @@
 import React, { useState } from 'react';
 import { useDispatch, useSelector } from "react-redux";
-
 import { makeChangeReview } from '../../store/reviews2';
-// import React, { useState } from "react";
-// import { useParams } from 'react-router-dom';
-// import { createReview } from '../../store/reviews2';
-// import { getOneSpot } from '../../store/spots2'
-// import './CreateReviewForm.css'
 import { useModal } from '../../context/Modal';
 import './editreview.css'
-// import '../CreateReviewModal/CreateReviewForm.css'
-// import * as reviewsAction from '../../store/reviews';
 
 
-function EditReviewForm({id, review}) {
+function EditReviewForm({ id, review }) {
 
     const { closeModal } = useModal();
     // onespot ?   :
@@ -28,38 +20,31 @@ function EditReviewForm({id, review}) {
     const handleSubmit = async (e) => {
         e.preventDefault()
         setErrors([]);
-        let reviewz = { id,  review }
+        let reviewz = { id, review }
         reviewz.review.stars = stars
         reviewz.review.review = reviewedit
-        // let reviewz = { id, review, stars, spotId: rev.spotId, type: true }
-        // console.log(reviewz)
         return dispatch(makeChangeReview(reviewz))
             .then(closeModal)
             .catch(async (res) => {
                 if (res.ok) {
                     const data = await res.json()
                     if (data.message) setErrors([data.message])
-                    // dispatch(getAllReviews(id))
-                }
-                // else setErrors(['something went wrong please try again!'])
-                // console.log(data, '!!!!!')
-                // else setErrors([res]);
 
-            }
-            )
+                }
+            })
         // .then(setShowModal(false))
     }
     // console.log({ idxx, review, stars })
     return (
 
         <form onSubmit={handleSubmit} id='createreviewform' >
-           <div id='signupheader'>
+            <div id='signupheader'>
                 <div id='loginexitbutt' onClick={() => closeModal()}>
                     <div>x</div>
                 </div>
                 <div id='signupheadertext'>
                     <div id="signupmainheader">{`Welcome to ${spot[id].name}!`}</div>
-                    <div id={!errors.length ? 'signupsubheader' : 'errorswap'}>{!errors.length ? 'Rate Your Stay!' : errors.map((error, idx) => <>{error}<br/></>)}</div>
+                    <div id={!errors.length ? 'signupsubheader' : 'errorswap'}>{!errors.length ? 'Rate Your Stay!' : errors.map((error, idx) => <>{error}<br /></>)}</div>
                 </div>
             </div>
