+++ frontend/src/components2/LoginFormModal/index.js
@@ -44,7 +44,6 @@ function LoginForm() {
         e.preventDefault();
         setErrors([]);
         return dispatch(sessionActions.login({ credential, password }))
-            // .then(() => setShowModal(false))
             .then(closeModal)
             .catch(async (res) => {
                 const data = await res.json()
@@ -56,15 +55,7 @@ function LoginForm() {
     }
     return (
         <form onSubmit={handleSubmit} id='loginform' >
-            {/* <div id="loginheader">
-                <div id='loginexitbutt' onClick={() => closeModal()}>
-                    <div>x</div>
-                </div>
-                <div>
-                    Login
-                </div>
-            </div> */}
-            
+
             <div id='signupheader'>
                 <div id='loginexitbutt' onClick={() => closeModal()}>
                     <div>x</div>
@@ -72,15 +63,9 @@ function LoginForm() {
                 <div id='signupheadertext'>
                     <div id="signupmainheader">Welcome to EarthRnR!</div>
                     <div id={!errors.length ? 'signupsubheader' : 'errorswap'}>{!errors.length ? ' Sign Up Today!' : errors.map((error, idx) => <>{error}</>)}</div>
-                    {/* {errors.map((error, idx) => <div className='errors' key={error + idx}>{error}</div>)} */}
                 </div>
             </div>
-            {/* <ul style={{ margin: '0' }}>
-                {errors.map((error, idx) => (
-                    <li className='errors' key={error + idx}>
-                        {error}</li>
-                ))}
-            </ul> */}
+
             <h3>Welcome to EarthRnR!</h3>
             <label className="loginlabel">
                 <input
