+++ frontend/src/components2/Navigation/Buttons/DemoButton.js
@@ -6,7 +6,6 @@ import './profile.css'
 
 const DemoButton = () => {
     const dispatch = useDispatch();
-    // const [errors, setErrors] = useState([]);
     const credential = 'D0rk5ter'
     const password = 'password'
     return <button id='dddemo' type="button" onClick={() => dispatch(login({ credential: credential, password: password }))}>Demo Site</button>
