+++ frontend/src/oldcomponents/DoubleSlider.js
